2011-07-16 14:28 UTC+0200 Viktor Szakats (harbour.01 syenar.hu)   * src/rtl/gtgui/gtgui.c     ! added cast to pacify msvc64 warning

+++ harbour/src/rtl/gtgui/gtgui.c
@@ -167,7 +167,7 @@ static int hb_gt_gui_optionPos( int id, int iType, PHB_ITEM pOptions )
    }
    if( iButton )
    {
-      int iOptions = hb_arrayLen( pOptions ), i;
+      int iOptions = ( int ) hb_arrayLen( pOptions ), i;
 
       for( i = 1; i <= iOptions; ++i )
       {
